Group some parsers inside namespaces.
