+++ src/AsciiDocNet.Tests/Unit/LabeledListItemTests.cs
@@ -1,3 +1,4 @@
+using AsciiDocNet.Lists;
 using Xunit;
 
 namespace AsciiDocNet.Tests.Unit
