+++ src/AsciiDocNet/Lists/LabeledList.cs
@@ -1,0 +1,98 @@
+using System.Collections.Generic;
+using System.Linq;
+
+namespace AsciiDocNet.Lists
+{
+	/// <summary>
+	/// A labeled list. Called also <see cref="https://docs.asciidoctor.org/asciidoc/latest/lists/description/">Description List in documentation</see>.
+	/// </summary>
+	/// <example>
+	/// CPU:: The brain of the computer.
+	/// Hard drive:: Permanent storage for operating system and/or user files.
+	/// </example>
+	/// <seealso cref="IElement" />
+	public class LabeledList : IElement
+	{
+		/// <summary>
+		/// Gets the items.
+		/// </summary>
+		/// <value>
+		/// The items.
+		/// </value>
+		public IList<LabeledListItem> Items { get; } = new List<LabeledListItem>();
+
+		/// <summary>
+		/// Gets or sets the parent element
+		/// </summary>
+		/// <value>
+		/// The parent.
+		/// </value>
+		public Container Parent { get; set; }
+
+		/// <summary>
+		/// Implements the operator ==.
+		/// </summary>
+		/// <param name="left">The left.</param>
+		/// <param name="right">The right.</param>
+		/// <returns>
+		/// The result of the operator.
+		/// </returns>
+		public static bool operator ==(LabeledList left, LabeledList right) => Equals(left, right);
+
+		/// <summary>
+		/// Implements the operator !=.
+		/// </summary>
+		/// <param name="left">The left.</param>
+		/// <param name="right">The right.</param>
+		/// <returns>
+		/// The result of the operator.
+		/// </returns>
+		public static bool operator !=(LabeledList left, LabeledList right) => !Equals(left, right);
+
+		/// <summary>
+		/// Determines whether the specified <see cref="object" />, is equal to this instance.
+		/// </summary>
+		/// <param name="obj">The <see cref="object" /> to compare with this instance.</param>
+		/// <returns>
+		///   <c>true</c> if the specified <see cref="object" /> is equal to this instance; otherwise, <c>false</c>.
+		/// </returns>
+		public override bool Equals(object obj)
+		{
+			if (ReferenceEquals(null, obj))
+				return false;
+			if (ReferenceEquals(this, obj))
+				return true;
+
+			return obj.GetType() == GetType() && Equals((LabeledList)obj);
+		}
+
+		/// <summary>
+		/// Returns a hash code for this instance.
+		/// </summary>
+		/// <returns>
+		/// A hash code for this instance, suitable for use in hashing algorithms and data structures like a hash table. 
+		/// </returns>
+		public override int GetHashCode() => Items.GetHashCode();
+
+		/// <summary>
+		/// Accepts a visitor to visit this element instance
+		/// </summary>
+		/// <typeparam name="TVisitor">The type of the visitor.</typeparam>
+		/// <param name="visitor">The visitor.</param>
+		/// <returns>
+		/// The visitor
+		/// </returns>
+		public TVisitor Accept<TVisitor>(TVisitor visitor) where TVisitor : IDocumentVisitor
+		{
+			visitor.VisitLabeledList(this);
+			return visitor;
+		}
+
+		/// <summary>
+		/// Determines whether the specified <see cref="LabeledList" />, is equal to this instance.
+		/// </summary>
+		/// <param name="other">The other.</param>
+		/// <returns>true if equal; otherwise, false</returns>
+		protected bool Equals(LabeledList other) => Items.SequenceEqual(other.Items);
+	}
+}
