+++ src/AsciiDocNet/Parsers/AttributeEntryParser.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Text.RegularExpressions;
+using AsciiDocNet.Attributes;
 
 namespace AsciiDocNet
 {
