+++ src/AsciiDocNet/Parsers/ElementAttributeParser.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
+using AsciiDocNet.Attributes;
 
 namespace AsciiDocNet
 {
