+++ src/AsciiDocNet/Parsers/LabeledListParser.cs
@@ -1,10 +1,11 @@
 using System;
 using System.Collections.Generic;
 using System.Text.RegularExpressions;
+using AsciiDocNet.Lists;
 
 namespace AsciiDocNet
 {
-    public class LabeledListParser : ProcessBufferParserBase
+	public class LabeledListParser : ProcessBufferParserBase
     {
         public override bool IsMatch(IDocumentReader reader, Container container, AttributeList attributes) =>
             PatternMatcher.LabeledListItem.IsMatch(reader.Line);
